user management system

+++ app/Http/Controllers/RolesController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\RoleCreateRequest;
-use App\Http\Requests\RoleUpdateRequest;
+use App\Models\Role;
 use App\Models\Features;
 use App\Models\Permissions;
-use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\RoleCreateRequest;
+use App\Http\Requests\RoleUpdateRequest;
 
 class RolesController extends Controller
 {
@@ -21,6 +22,10 @@ class RolesController extends Controller
 
     public function store(RoleCreateRequest $request)
     {
+        if(Auth::user()->roles->id != 1){
+            return back()->with('message', 'Unauthorize');
+        }
+        
         $role = new Role();
         $role->name = $request->role_name;
         $role->save();
@@ -33,6 +38,10 @@ class RolesController extends Controller
 
     public function edit(Role $role)
     {
+        if(Auth::user()->roles->id != 1){
+            return back()->with('message', 'Unauthorize');
+        }
+
         $check_permission = [];
         foreach ($role->permissions as $permission) {
             $check_permission[] = $permission->id;
@@ -44,6 +53,10 @@ class RolesController extends Controller
 
     public function update(RoleUpdateRequest $request,Role $role)
     {
+        if(Auth::user()->roles->id != 1){
+            return back()->with('message', 'Unauthorize');
+        }
+
         $role->name = $request->role_name;
         $role->save();
         $permission = $request->input('user_management');
@@ -55,6 +68,10 @@ class RolesController extends Controller
 
     public function delete(Role $role)
     {
+        if(Auth::user()->roles->id != 1){
+            return back()->with('message', 'Unauthorize');
+        }
+        
         $role->delete();
         $permission_id = $role->permissions()->pluck("permissions_id");
         $role->permissions()->detach($permission_id);
